+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- New `overwrite_exiting` option on `install::Spec`. Default is false, by
+  setting it to true the installer will overwrite existing executables.
+
 # Version 0.3.0 (2024-4-9)
 
 ### Changed